Commit files for initial DG implementation
This differs from Rogers-Ricci in that it has reduced gradients (wider source function) and reduced nonlinearity (by linearizing exp(y) to 1 + y).  This runs and produces something that has the appearance of a turbulent state near the end (i.e. it does not seem to be a laminar flow: not smooth, not symmetric).